LIBITD-1940. Removed Gateway abstract class
It seems unlikely that there will ever be any other implementation
of the "Gateway" abstract class than "DbGateway", so removing it
as an unnecessary complication.

https://issues.umd.edu/browse/LIBITD-1940
