+++ patsy/commands/schema.py
@@ -1,6 +1,6 @@
 import argparse
 import patsy.core.command
-from patsy.core.gateway import Gateway
+from patsy.core.db_gateway import DbGateway
 from patsy.database import create_schema
 
 
@@ -16,5 +16,6 @@ def configure_cli(subparsers) -> None:  # type: ignore
 
 
 class Command(patsy.core.command.Command):
-    def __call__(self, args: argparse.Namespace, gateway: Gateway) -> str:
+    def __call__(self, args: argparse.Namespace, gateway: DbGateway) -> str:
         create_schema(args)
+        return "Schema created"
